requirements.txt
Django>=5.0
PyMySQL>=1.1.0
python-decouple>=3.8
Pillow>=10.0
requests>=2.31
whitenoise>=6.0
gunicorn>=21.0
